WIP: Add syncwaves to applications, namespaces, subscriptions, projects
The argo sync-wave logic would be to determine the first application in
a namespace, and bring up the namespace before that with the
namespaceOffset. Same would go for the project, determine the first
application in a project.
For subscription, the first application in the subscription's target namespace
would be the reference, and the subscription would be brought up before
that, by the offset.